# Parser optimizations