--- a/src/generated/ldb_animationcelldata.cpp
+++ b/src/generated/ldb_animationcelldata.cpp
@@ -18,79 +18,90 @@
 
 template <>
 char const* const Struct<RPG::AnimationCellData>::name = "AnimationCellData";
+static TypedField<RPG::AnimationCellData, int32_t> static_valid(
+	&RPG::AnimationCellData::valid,
+	LDB_Reader::ChunkAnimationCellData::valid,
+	"valid",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_cell_id(
+	&RPG::AnimationCellData::cell_id,
+	LDB_Reader::ChunkAnimationCellData::cell_id,
+	"cell_id",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_x(
+	&RPG::AnimationCellData::x,
+	LDB_Reader::ChunkAnimationCellData::x,
+	"x",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_y(
+	&RPG::AnimationCellData::y,
+	LDB_Reader::ChunkAnimationCellData::y,
+	"y",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_zoom(
+	&RPG::AnimationCellData::zoom,
+	LDB_Reader::ChunkAnimationCellData::zoom,
+	"zoom",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_tone_red(
+	&RPG::AnimationCellData::tone_red,
+	LDB_Reader::ChunkAnimationCellData::tone_red,
+	"tone_red",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_tone_green(
+	&RPG::AnimationCellData::tone_green,
+	LDB_Reader::ChunkAnimationCellData::tone_green,
+	"tone_green",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_tone_blue(
+	&RPG::AnimationCellData::tone_blue,
+	LDB_Reader::ChunkAnimationCellData::tone_blue,
+	"tone_blue",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_tone_gray(
+	&RPG::AnimationCellData::tone_gray,
+	LDB_Reader::ChunkAnimationCellData::tone_gray,
+	"tone_gray",
+	0,
+	0
+);
+static TypedField<RPG::AnimationCellData, int32_t> static_transparency(
+	&RPG::AnimationCellData::transparency,
+	LDB_Reader::ChunkAnimationCellData::transparency,
+	"transparency",
+	0,
+	0
+);
+
 
 template <>
 Field<RPG::AnimationCellData> const* Struct<RPG::AnimationCellData>::fields[] = {
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::valid,
-		LDB_Reader::ChunkAnimationCellData::valid,
-		"valid",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::cell_id,
-		LDB_Reader::ChunkAnimationCellData::cell_id,
-		"cell_id",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::x,
-		LDB_Reader::ChunkAnimationCellData::x,
-		"x",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::y,
-		LDB_Reader::ChunkAnimationCellData::y,
-		"y",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::zoom,
-		LDB_Reader::ChunkAnimationCellData::zoom,
-		"zoom",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::tone_red,
-		LDB_Reader::ChunkAnimationCellData::tone_red,
-		"tone_red",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::tone_green,
-		LDB_Reader::ChunkAnimationCellData::tone_green,
-		"tone_green",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::tone_blue,
-		LDB_Reader::ChunkAnimationCellData::tone_blue,
-		"tone_blue",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::tone_gray,
-		LDB_Reader::ChunkAnimationCellData::tone_gray,
-		"tone_gray",
-		0,
-		0
-	),
-	new TypedField<RPG::AnimationCellData, int32_t>(
-		&RPG::AnimationCellData::transparency,
-		LDB_Reader::ChunkAnimationCellData::transparency,
-		"transparency",
-		0,
-		0
-	),
+	&static_valid,
+	&static_cell_id,
+	&static_x,
+	&static_y,
+	&static_zoom,
+	&static_tone_red,
+	&static_tone_green,
+	&static_tone_blue,
+	&static_tone_gray,
+	&static_transparency,
 	NULL
 };