… CMakeLists.txt
… bench/readldb.cpp
… src/generated/ldb_battlecommands.cpp
… src/generated/ldb_battleranimation.cpp
… src/generated/ldb_battleranimationdata.cpp
… src/generated/ldb_battleranimationextension.cpp
… src/generated/ldb_chipset.cpp
… src/generated/ldb_class.cpp
… src/generated/ldb_commonevent.cpp
… src/generated/ldb_database.cpp
… src/generated/ldb_enemy.cpp
… src/generated/ldb_enemyaction.cpp
… src/generated/ldb_item.cpp
… src/generated/ldb_itemanimation.cpp
… src/generated/ldb_music.cpp
… src/generated/ldb_skill.cpp
… src/generated/ldb_sound.cpp
… src/generated/ldb_state.cpp
… src/generated/ldb_switch.cpp
… src/generated/ldb_system.cpp
… src/generated/ldb_terms.cpp
… src/generated/ldb_terrain.cpp
… src/generated/ldb_testbattler.cpp
… src/generated/ldb_troop.cpp
… src/generated/ldb_troopmember.cpp
… src/generated/ldb_trooppage.cpp
… src/generated/ldb_trooppagecondition.cpp
… src/generated/ldb_variable.cpp
… src/generated/lmt_encounter.cpp
… src/generated/lmt_mapinfo.cpp