`bench_readldb` on master:

| Game | Run Time | Mem Usage | # malloc | # lseek | # read |
| -- | -- | -- | -- | -- | -- |
| Heroes Realm | 2.69s | 289.4MB | 7,399,774 | 3,481,059 | 20,077 |
| HH3 | 1.29s | 134.3MB | 3,4724,652 | 1,663,829 | 13,244 |
| 2k3 New Project | 0.077s | 8.9MB | 8,979 | 146,217 | 156 |

`bench_readldb` on  "Cache the file offset"

| Game | Run Time | # lseek | # read | SpeedUp |
| -- | -- | -- | -- | -- |
| Heroes Realm | 1.24s | 41165 | 20777 | 2.17x |
| HH3 | 0.58s | 25589 | 13244 | 2.24x |
| 2k3 New Project | 0.015s | 199 | 156 | 5.13x |

`bench_readldb` on "More efficient loading of EventCommand parameters"

| Game | Run Time | # malloc | SpeedUp | Cum SpeedUp |
| -- | -- | -- | -- | -- |
| Heroes Realm | 1.15s | 3,395,008 | 1.08x | 2.34x |
| HH3 | 0.54s | 1,620,846 | 1.07x | 2.39x |
| 2k3 New Project | 0.015s | 8979 | 1.0x | 5.13x |

`bench_readldb` on  "Optimize out calls to seekg()"

| Game | Run Time | # lseek | # read | SpeedUp | Cum SpeedUp |
| -- | -- | -- | -- | -- | -- |
| Heroes Realm | 1.13s | 1 | 3446 | 1.01x | 2.38x |
| HH3 | 0.52s | 1 | 1969 | 1.04x | 2.48x |
| 2k3 New Project | 0.015s | 1 | 60 | 1.0x | 5.13x |