@fmatthew5876 2019-09-07
The first commit "Cache the file offset"

I tested it with HH3's database which is 16MB and has a high proportion of battle events.

Before:
```
Info: Loading RPG_RT database `../../RPG_RT.ldb'...

 Performance counter stats for './copydb/copydb ../../RPG_RT.ldb copy.ldb':

          1,784.33 msec task-clock                #    0.990 CPUs utilized          
                 2      context-switches          #    0.001 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            32,220      page-faults               #    0.018 M/sec                  
     6,703,250,122      cycles                    #    3.757 GHz                    
    11,104,754,123      instructions              #    1.66  insn per cycle         
     2,550,663,441      branches                  # 1429.482 M/sec                  
        14,320,378      branch-misses             #    0.56% of all branches        

       1.801442636 seconds time elapsed

       1.492500000 seconds user
       0.292098000 seconds sys
```

After:
```
Info: Loading RPG_RT database `../../RPG_RT.ldb'...

 Performance counter stats for './copydb/copydb ../../RPG_RT.ldb copy.ldb':

          1,146.98 msec task-clock                #    0.979 CPUs utilized          
                 5      context-switches          #    0.004 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            32,219      page-faults               #    0.028 M/sec                  
     4,314,516,454      cycles                    #    3.762 GHz                    
    10,281,004,010      instructions              #    2.38  insn per cycle         
     2,382,100,523      branches                  # 2076.842 M/sec                  
        11,127,746      branch-misses             #    0.47% of all branches        

       1.171960337 seconds time elapsed

       1.075344000 seconds user
       0.071956000 seconds sys
```