@fmatthew5876 2019-09-08
The next one "More efficient loading of EventCommand parameters"


Heroes Realm Loading only before:
```
 Performance counter stats for './copydb/copydb /home/matt/Documents/RPG Maker 2003/HerosRealmRTP/RPG_RT.ldb copy.ldb':

          1,182.84 msec task-clock                #    1.000 CPUs utilized          
                 3      context-switches          #    0.003 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            70,992      page-faults               #    0.060 M/sec                  
     4,385,285,132      cycles                    #    3.707 GHz                    
     9,209,959,474      instructions              #    2.10  insn per cycle         
     1,918,129,321      branches                  # 1621.633 M/sec                  
         8,011,706      branch-misses             #    0.42% of all branches        

       1.183190699 seconds time elapsed

       1.078858000 seconds user
       0.104276000 seconds sys


 Performance counter stats for './copydb/copydb /home/matt/Documents/RPG Maker 2003/HerosRealmRTP/RPG_RT.ldb copy.ldb':

         7,399,774      probe_libc:malloc                                           

       6.967116686 seconds time elapsed

       3.425364000 seconds user
       3.541410000 seconds sys
```

After
```


 Performance counter stats for './copydb/copydb /home/matt/Documents/RPG Maker 2003/HerosRealmRTP/RPG_RT.ldb copy.ldb':

          1,053.10 msec task-clock                #    1.000 CPUs utilized          
                 2      context-switches          #    0.002 K/sec                  
                 0      cpu-migrations            #    0.000 K/sec                  
            71,033      page-faults               #    0.067 M/sec                  
     3,974,947,922      cycles                    #    3.775 GHz                    
     8,348,157,566      instructions              #    2.10  insn per cycle         
     1,717,461,359      branches                  # 1630.858 M/sec                  
         9,509,200      branch-misses             #    0.55% of all branches        

       1.053524610 seconds time elapsed

       0.949360000 seconds user
       0.104149000 seconds sys

 Performance counter stats for './copydb/copydb /home/matt/Documents/RPG Maker 2003/HerosRealmRTP/RPG_RT.ldb copy.ldb':

         3,477,268      probe_libc:malloc                                           

       3.809424821 seconds time elapsed

       2.051271000 seconds user
       1.755376000 seconds sys
```