@fmatthew5876 2019-09-10
Reading an ldb file on "Optimize out calls to seekg()"

| Game | Run Time | # lseek | # read | SpeedUp | Cum SpeedUp |
| -- | -- | -- | -- | -- | -- |
| Heroes Realm | 2.46s | 1 | 3446 | 2% | 66% |
| HH3 | 1.17s | 1 | 1969 | 3% | 60% |
| 2k3 New Project | 0.028s | 1 | 60 | 0% | 217% |